refactor: rename tokenize_sig.py
