+++ src/suricatals/langserver.py
@@ -33,7 +33,7 @@ import queue
 from suricatals.signature_parser import SuricataFile
 from suricatals.signature_validator import TestRules
 from suricatals.suricata_command import SuriCmd
-from suricatals.tokenize_sig import SuricataSemanticTokenParser
+from suricatals.signature_tokenizer import SuricataSemanticTokenParser
 
 
 from pygls.lsp.server import LanguageServer
